# api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cat Activities Monitor API",
    "description": "\n    A comprehensive FastAPI-based system for monitoring cats using YOLO object detection and activity analysis.    ",
    "license": {
      "name": "MIT License",
      "url": "https://github.com/biancarosa/cat-activities-monitor/blob/main/LICENSE"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://your-domain.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/cameras": {
      "get": {
        "tags": [
          "cameras"
        ],
        "summary": "List Camera Sources",
        "description": "Get a list of all configured camera/image sources with their settings and status.",
        "operationId": "list_cameras_cameras_get",
        "responses": {
          "200": {
            "description": "Camera sources retrieved successfully",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "cameras": [
                    {
                      "name": "living_room",
                      "url": "http://camera1.local/snapshot.jpg",
                      "interval_seconds": 30,
                      "enabled": true
                    },
                    {
                      "name": "kitchen",
                      "url": "http://camera2.local/snapshot.jpg",
                      "interval_seconds": 60,
                      "enabled": false
                    }
                  ],
                  "total": 2,
                  "enabled": 1
                }
              }
            }
          },
          "404": {
            "description": "No configuration loaded"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/cameras/{camera_name}/fetch": {
      "post": {
        "tags": [
          "cameras"
        ],
        "summary": "Fetch Specific Camera Image",
        "description": "Fetch and analyze a specific camera image by name. This will download the image, run YOLO detection, and return the results.",
        "operationId": "fetch_camera_image_cameras__camera_name__fetch_post",
        "parameters": [
          {
            "name": "camera_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Camera Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Camera image fetched and analyzed successfully",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "message": "Successfully fetched 'living_room'",
                  "camera_name": "living_room",
                  "url": "http://camera1.local/snapshot.jpg",
                  "success": true,
                  "detection": {
                    "detected": true,
                    "confidence": 0.89,
                    "count": 2,
                    "detections": [
                      {
                        "class_id": 15,
                        "confidence": 0.89,
                        "bounding_box": {
                          "x1": 100.5,
                          "y1": 150.2,
                          "x2": 300.8,
                          "y2": 400.1,
                          "width": 200.3,
                          "height": 249.9
                        }
                      }
                    ],
                    "total_animals": 2,
                    "activities": []
                  }
                }
              }
            }
          },
          "404": {
            "description": "Camera source not found or no configuration loaded"
          },
          "500": {
            "description": "Failed to fetch or analyze camera image"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cameras/fetch-all": {
      "post": {
        "tags": [
          "cameras"
        ],
        "summary": "Fetch All Camera Images",
        "description": "Fetch and analyze images from all enabled camera sources concurrently.",
        "operationId": "fetch_all_cameras_cameras_fetch_all_post",
        "responses": {
          "200": {
            "description": "Camera images fetch completed",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "message": "Fetch completed: 3/4 successful",
                  "total_cameras": 4,
                  "enabled_cameras": 4,
                  "successful": 3,
                  "failed": 1,
                  "results": [
                    {
                      "camera_name": "living_room",
                      "success": true,
                      "detection": {
                        "detected": true,
                        "count": 2
                      }
                    },
                    {
                      "camera_name": "kitchen",
                      "success": false,
                      "error": "Connection timeout"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No configuration loaded"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}